requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-pytest==2.13.2
webdriver-manager==4.0.1
openpyxl==3.1.2
requests==2.32.4
Pillow==10.1.0